Modifying delete for hfc

+++ build/web/SystemAdmin/master_lookup_delete.jsp
@@ -4,6 +4,7 @@
     Author     : user
 --%>
 
+<%@page import="java.util.ArrayList"%>
 <%@page import="Config.Config"%>
 <%@page import="dBConn.Conn"%>
 <%@page import="main.RMIConnector"%>
@@ -18,13 +19,28 @@
     String id = request.getParameter("masterCode");
 
     RMIConnector rmic = new RMIConnector();
-    String sql = "DELETE FROM adm_lookup_master WHERE master_reference_code = '" + id + "' ";
+    
+    String sqlSelect = "Select master_reference_code from adm_lookup_detail where master_reference_code = '"+id+"' LIMIT 1 ";
+    
+    ArrayList<ArrayList<String>> dataUse = conn.getData(sqlSelect);
+    
+    if(dataUse.size() > 0){
+    
+        out.print("You can't delete this item because it is referrenced by lookup detail.");
+        
+    }else{
+    
+        String sql = "DELETE FROM adm_lookup_master WHERE master_reference_code = '" + id + "' ";
 
-    boolean status = rmic.setQuerySQL(conn.HOST, conn.PORT, sql);
+        boolean status = rmic.setQuerySQL(conn.HOST, conn.PORT, sql);
 
-    if (status == true) {
-        out.print("Success");
-    } else {
-        out.print("Failed");
+        if (status == true) {
+            out.print("Success");
+        } else {
+            out.print("Failed");
+        }
+        
     }
+    
+    
 %>
